+++ client/src/constant.ts
@@ -1,5 +1,3 @@
-import { imageHandler } from '@/components/QuillEditBox/imageHandler';
-
 export const linkList = [
   { name: '게시판', href: '/board' },
   { name: 'Q&A', href: '/qna' },
@@ -79,29 +77,6 @@ export const postList = [
     view: 5,
   },
 ];
-
-export const quillModules = {
-  toolbar: {
-    container: [
-      [{ header: '1' }, { header: '2' }, { font: [] }],
-      [{ size: [] }],
-      ['bold', 'italic', 'underline', 'strike', 'blockquote'],
-      [
-        { list: 'ordered' },
-        { list: 'bullet' },
-        { indent: '-1' },
-        { indent: '+1' },
-      ],
-      ['link', 'image'],
-    ],
-    handlers: {
-      image: imageHandler,
-    },
-  },
-  clipboard: {
-    matchVisual: false,
-  },
-};
 
 export const quillFormats = [
   'header',
